+++ dclean/__init__.py
@@ -5,5 +5,5 @@ Import the main entry point:
 """
 from .core import Data
 
-__version__ = "0.1.4"
+__version__ = "0.1.5"
 __all__ = ["Data"]